Connected our libGDX code with Board and Tile
First attempt at implementing game logic, currently we only check if there is a tile of x type in the tile we wish to move to. If there is a wall in this tile, the piece does not move.

Future implementation will allow movement to the tile, but not in between tiles which have a wall separating them.