Board is now read and stored as TileObjects
First implementation of TileObjects with room for more advanced logic than our previous Tile enum.

!!This breaks all the tests!! - rewriting them now

Same bugs as before, if there are two tiles on the same spot, this will only save the one last read. In other words a TileObject is saved over the current one in the same position if another is read for the same location on the board